현재까지 작업물

+++ src/main/java/hocheoltech/boos/service/BoardService.java
@@ -1,6 +1,8 @@
 package hocheoltech.boos.service;
 
 import hocheoltech.boos.domain.Board;
+import hocheoltech.boos.exception.DuplicateMemberIdException;
+import hocheoltech.boos.exception.NotExistBoard;
 import hocheoltech.boos.repository.BoardRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -18,19 +20,25 @@ public class BoardService {
 
     // 게시글 등록
     public Board createBoard(Board board){
-//        boardMapper.createBoard(board);
-        return boardRepository.save(board);
+        Board savedBoard = boardRepository.save(board);
+        return savedBoard;
     }
 
     // 게시글 삭제
     public void deleteBoard(long seq){
-//        boardMapper.deleteBoard(seq);
+        if(!boardRepository.existsById(seq)){
+            throw new NotExistBoard("There are no board matching the seq : "+seq);
+        }
         boardRepository.deleteById(seq);
     }
 
     // 게시글 수정
-    public void updateBoard(Board board){
-        //boardMapper.updateBoard(board);
+    public void updateBoard(Long seq, Board board){
+        if(!boardRepository.existsById(seq)){
+            throw new NotExistBoard("There are no board matching the seq : "+seq);
+        }
+        Board existBoard = boardRepository.findById(seq).get();
+        existBoard.updateBoard(board.getTitle(),board.getContent(),board.getCategory()); // 영속성 프레임워크에 의한 자동 update
     }
 
     // 게시판리스트 조회
@@ -39,8 +47,11 @@ public class BoardService {
     }
 
     // 게시판 상세 조회
-    public Optional<Board> getBoardDetail(long seq){
-        return boardRepository.findById(seq);
+    public Board getBoardDetail(long seq){
+        if(!boardRepository.existsById(seq)){
+            throw new NotExistBoard("There are no board matching the seq : "+seq);
+        }
+        return boardRepository.findById(seq).get();
     }
 
 }
